changed handling for "rememberCookieConsentGiven"
If there is no Google Consent Mode initialization at all in place (which is the usual setup if one decides against Google tagging and instead wants to use Matomo; maybe without any consent promt), the template API call "isConsentGranted('analytics_storage')" will always return "true". 

This may become a problem because "rememberCookieConsentGiven" will always be executed in such setups with the current template code. This results in an unwanted mtm_cookie_consent cookie. 

It might be a solution to bind the usage of the Google Consent Mode state for setting the rememberCookieConsentGiven option to the useGoogleConsentAPI setting: 

if (((data.useGoogleConsentAPI == true) && isConsentGranted('analytics_storage')) || useRememberCookieConsentGiven == true || useRememberCookieConsentGiven == "true") {...

This would allow settings like "disableBrowserFeatureDetection" to work, too. Without "dismantling" the Consent API, "rememberCookieConsentGiven" overrides the option (I added in my forked version) and plugins and resolution will always be part of the data capturing in the browser (problematic in Germany due to TTDSG).